refactor problems
