+++ problems/n_dim/square.py
