+++ problems/one_dim/polynomial_1d.py
