+++ sao/solvers/method_of_moving_asymptotes.py
@@ -7,7 +7,7 @@ from sao.solvers.primal_dual_interior_point import pdip
 
 
 def mma(problem, x0=None, move=0.2, xmin=0.0, xmax=1.0, stop_tol=1e-6):
-    int_variable = MMA(xmin=xmin, xmax=xmax)
+    int_variable = MMA(x_min=xmin, x_max=xmax)
     approx = Taylor1(int_variable)
     lim1 = Bounds(problem.x_min, problem.x_max)
     lim2 = MoveLimit(move * (problem.x_max - problem.x_min))
@@ -25,6 +25,6 @@ def mma(problem, x0=None, move=0.2, xmin=0.0, xmax=1.0, stop_tol=1e-6):
         df = problem.dg(x)
         print(iter, ":  ", f[0], x)
         sub_problem.build(x, f, df)
-        x[:] = pdip(sub_problem)
+        x[:] = pdip(sub_problem)[0]
     f = problem.g(x)
     return x, f[0]
